Expression parsing refactor
