+++ src/parsing_type.c
@@ -99,267 +99,3 @@ match_t *parse_type(char *line) {
 
 	return match;
 }
-
-static void parse_variable_expression(char *line, char **name, char *is_pointer) {
-	unsigned int start_index = 0;
-	unsigned int end_index   = 0;
-	int index                = 0;
-	int length               = strlen(line);
-	char c;
-
-	//Pre incrementing / decrementing
-	SKIP_WHITESPACES
-	if(strstr(line + index, "++") == line + index || strstr(line + index, "--") == line + index) {
-		//Skip operation
-		index++;	
-	} else if(index >= 0) {
-		index--;
-		c = line[index];
-	}
-
-	//Check referencing
-	do {
-		index++;
-		SKIP_WHITESPACES
-
-		if(c == '*') (*is_pointer)--;
-		if(c == '&') (*is_pointer)++;
-	} while(c == '*' || c == '&');
-
-	SKIP_WHITESPACES
-
-	start_index = index;
-
-	//Until not alphanumeric
-	while(is_alphanumeric(c = line[index]) && index < length) {
-		index++;
-	}
-
-	end_index = index;
-
-	*name = strsubstr(line, start_index, end_index - start_index);
-
-}
-
-static unsigned char is_type_identifier(char c) {
-	return c == 'u' || c == 'l' || c == 'f' || c == 'U' || c == 'L' || c == 'F' || c == 'd' || c == 'D'; 
-}
-
-type_t get_expression_type(char *line, int line_index, scope_t *scope, arraylist_t *undeclared_variables, arraylist_t *undeclared_functions, arraylist_t *invalid_params) { //TODO rename to parse_expression
-	type_t type;
-	function_t *function     = NULL;
-	function_t *function_dec = NULL;
-	field_t *variable_dec    = NULL;
-	char *variable_name      = NULL;
-	char *tmp                = NULL;
-	char *type_identifier    = NULL; 
-	char is_pointer          = 0;
-	unsigned int start_index = 0;
-	unsigned int end_index   = 0;
-	int index                = 0;
-	unsigned int sub_index   = 0;
-	int length               = strlen(line);
-	int type_length;
-	char c;
-
-	type.name = strduplicate("NULL");
-	type.is_pointer = 0;
-	type.is_literal = 0;
-
-	SKIP_WHITESPACES
-
-	if(c == '(') {
-		//Parenthesis -> cast
-
-		index++;
-
-		SKIP_WHITESPACES
-
-		start_index = index;
-
-		while((c = line[index]) != ')' && index < length) index++;
-
-		if(c != ')') { //No closing parenthesis, syntax error
-			return type;
-		}
-
-		//Skip whitespaces backwards to find end of type
-		while(is_whitespace(c = line[index]) && index > 0) { index--; }
-
-		end_index = index;
-
-		free(type.name);
-		type_length     = end_index - start_index;
-		type.name       = strsubstr(line, start_index, type_length);
-		type.is_pointer = strcount(type.name, '*');
-		sub_index       = strcountuntil(type.name, '*', 1, 1);
-
-		//Remove stars
-		type.name[type_length - sub_index] = '\0';
-
-	} else if(c == '\'' && strlastindexof(line, '\'') != index) {
-		free(type.name);
-		type.name       = strduplicate("char");
-		type.is_pointer = 0;
-		type.is_literal = 1;
-	} else if(c == '"' && strlastindexof(line, '"') != index) {
-		free(type.name);
-		type.name       = strduplicate("char");
-		type.is_pointer = 1;
-		type.is_literal = 1;
-	} else if(undeclared_variables != NULL) {
-
-		if(!is_digit(c) && c != '.') {
-
-			function = parse_function_call(line_index, line + index);
-			if(function != NULL) {
-				if(scope == NULL)
-					arraylist_add(undeclared_functions, function);
-				else {
-
-					function_dec = find_function(scope, function->name, 0);
-					if(function_dec != NULL && (function_dec->line < line_index || find_function_prototype(get_root_scope(scope), function->name) != NULL)) {
-						free(type.name);
-						type.name = strduplicate(function_dec->return_type.name);
-						type.is_pointer = function_dec->return_type.is_pointer;
-						type.is_literal = 0;
-						check_function_call_parameters(scope, function, function_dec, line_index, line, undeclared_variables, undeclared_functions, invalid_params);
-						function_free(function);
-					} else {
-						arraylist_add(undeclared_functions, function);
-						check_function_call_parameters(scope, function, function_dec, line_index, line, undeclared_variables, undeclared_functions, invalid_params);
-					}
-				}
-
-			} else {
-				parse_variable_expression(line + index, &variable_name, &is_pointer);
-				if(variable_name != NULL) {
-
-					if(scope == NULL || (variable_dec = find_variable(scope, variable_name)) == NULL)
-						arraylist_add(undeclared_variables, variable_name);
-					else {
-						type.is_pointer = variable_dec->type.is_pointer + is_pointer;
-						if(type.is_pointer < 0) { //More dereferencing than allowed
-							type.is_pointer = 0;
-							type.is_literal = 0;
-						} else {
-							type.name = strduplicate(variable_dec->type.name);
-							type.is_literal = 0;
-						}
-					}
-				}
-			}
-		} else {
-			//TODO numbers
-			//Starts with 0x -> integer
-			//Integer -> only digits
-			//ends with u or U -> unsigned
-			//follows L or l -> long (while, max 2)
-
-			//Double -> only digits, one dot, can end with 'd'
-			//Can start with dot
-			//follows L or l -> long (max 1)
-			//float -> only digits, one dot, ends with 'f'
-
-			//Get all digits
-			while(is_digit(c = line[index]) && index < length) {
-				index++;
-			}
-
-			if(c == '.') { //Float or double
-				c = line[++index];
-				if(is_digit(c)) {
-					while(is_digit(c = line[index]) && index < length) {
-						index++;
-					}
-
-					if(index >= length) {
-						type_identifier = "double";
-					} else if(is_type_identifier(c)) {
-						switch(c) {
-						case 'f':
-						case 'F':
-							type_identifier = "float";
-							break;
-						case 'd':
-						case 'D':
-							type_identifier = "double";
-							break;
-						case 'l':
-						case 'L':
-							type_identifier = "long double";
-							break;
-						}
-
-					}
-
-					free(type.name);
-					type.name = strduplicate(type_identifier);
-				}
-			} else if(is_type_identifier(c)) {
-
-				while(is_type_identifier(c) && index < length) {
-
-					switch(c) {
-						case 'u':
-						case 'U':
-							type_identifier = !strcmp(type.name, "NULL") ? "unsigned" : "unsigned ";
-							break;
-						case 'l':
-						case 'L':
-							type_identifier = !strcmp(type.name, "NULL") ? "long" : " long";
-							break;
-						case 'f':
-						case 'F':
-							type_identifier = !strcmp(type.name, "NULL") ? "float" : " float";
-							break;
-						case 'd':
-						case 'D':
-							type_identifier = !strcmp(type.name, "NULL") ? "double" : " double";
-							break;
-					}
-
-					if(!strcmp(type.name, "NULL")) {
-						free(type.name);
-						type.name = strduplicate(type_identifier);
-					} else {
-						
-						//If unsigned, float or double already set, wrong
-						if(!strcmp(type_identifier, "unsigned "))
-							tmp = strconcat(type_identifier, type.name);
-						else
-							tmp = strconcat(type.name, type_identifier);
-						free(type.name);
-						type.name = tmp;
-					}
-
-					if(index < length)
-						c = line[++index];
-				}
-
-				if(index < length && c != ';') { //Syntax error
-					free(type.name);
-					type.name = strduplicate("NULL");
-				} else if(!strcmp(type.name, "unsigned")) {
-					//If just "unsigned", add "int" at the end
-					tmp = strconcat(type.name, " int");
-					free(type.name);
-					type.name = tmp;
-				}
-
-			} else {
-				type.name = strduplicate("int");
-				type.is_literal = 1;
-				type.is_pointer = 0;
-			}
-		}
-
-	}
-
-
-	//TODO operations with operators (+, -, ..., &&, ||)
-
-	//TODO conditional expressions and loops
-	//if, else if, for, switch, case, while
-	return type;
-}
