+++ src/rules/undeclared_function.c
@@ -1,5 +1,6 @@
 #include "rules/undeclared_function.h"
 #include "parsing_type.h"
+#include "parsing_expressions.h"
 #include "arraylist.h"
 #include "display.h"
 
@@ -33,7 +34,7 @@ unsigned int check_undeclared_functions(scope_t *root_scope, arraylist_t *file) 
 			invalid_params       = arraylist_init(ARRAYLIST_DEFAULT_CAPACITY);
 
 			line = arraylist_get(file, i);
-			type = get_expression_type(line, i, scope, undeclared_variables, undeclared_functions, invalid_params); //Check type equals
+			type = parse_expression(line, i, scope, undeclared_variables, undeclared_functions, invalid_params); //Check type equals
 			for(size_t j = 0 ; j < undeclared_functions->size ; j++) {
 				function = arraylist_get(undeclared_functions, j);
 				message = strconcat("Undeclared function: ", function->name);
